+++ src/main/kotlin/no/nsd/qddt/repository/PublicationStatusRepository.kt
@@ -12,7 +12,7 @@ import java.util.*
  */
 
 @RepositoryRestResource(path = "publicationstatus", exported = false , itemResourceRel = "PublicationStatus", excerptProjection = PublicationStatusListe::class)
-interface PublicationStatusRepository : JpaRepository<PublicationStatus, Long> {
+interface PublicationStatusRepository : JpaRepository<PublicationStatus, Int> {
 
 
 }
